Adding delete function, delete confrimation, and deletion routes

+++ resources/views/siswa/index.blade.php
@@ -35,7 +35,7 @@
                         <td>{{$siswa->address}}</td>
                         <td>
                           <a href="/siswa/{{$siswa->id}}/edit" class="btn btn-warning btn-sm">Edit</a>
-                          <a href="/siswa/{{$siswa->id}}/delete" class="btn btn-danger btn-sm">Delete</a>
+                          <a href="/siswa/{{$siswa->id}}/delete" class="btn btn-danger btn-sm" onclick="return confirm('Delete Confirmation')">Delete</a>
                         </td>
                     </tr>
                     @endforeach
